feat(POST): upload simple text files
Experimenting with POST requests and understanding how they should work.
This is a hacky job because it very unelegantly tries to parse the body
of the request in order to get Content-Type and Content-Disposition
fields, please fix later.
Right now it assumes there is only one block of content in the request
body and treats that as the file to write, which is probably incorrect.