feat(server): implement request/response queues to better serve clients
A major problem encountered during our last testing was how a single
client could hold up our entire server forever due to the way we recv()
bytes.
Originally, we block at epoll_wait() until a connection comes in,
in which we recv() bytes in a loop from said connection until a
verifiably (in)valid Request was formed.
If said connection was slow, it would hold up our server when we could
be serving other clients in the meantime instead.

The event loop has been restructured such that when we block at
epoll_wait() and receive an incoming stream of bytes, we recv() once
into a unique buffer belonging to each connection.
The processing of a stream of bytes into a Request is now decoupled
from the recv().
Once we have a valid Request, we can add it to a queue of Requests
to be consumed. By going through the queue, we can generate
corresponding Responses which we currently place in a separate queue
of Responses that we can then go through to send a Response to each client.

This seems to fix one client potentially holding up our server and also
seems to have increased our transactions/sec.