feat(http): add MediaType class that maps file ext to MIME types
Add a MediaType class in the Server class that maps file extensions to
MIME types. For now it reads from a file and maps accordingly (no
format checking). By right there should be a map within code which
would be replaced by user-defined config, depending on the existence of
the mime.types file.

The construction of Content-Type header will use this map to add the
appropriate media-type value based on the file extension within the
file name.