feat(string): add functions to match string based on predicate
consumeIf() is passed in a Predicate function object that can be
initialized to a string indicating the characters to be matched.

consumeIfUntil() consumes characters up to an expected "terminator", but
also signals an error if an unexpected character that does not match the
Predicate is found.

Unsure if it is better to return true/false error codes or throw
exceptions, this should be fixed later when we put this to practical
use.