feat(POST): parse multipart/form-data and form-urlencoded
The only 2 content types handled so far are form-urlencoded and
multipart/form-data, where for urlencoded the key value pairs in the
body are read and stored into a map for later processing. For
multipart/form-data, all the parts are parsed through and files are
uploaded to the server. The simplest form of error detection right now
is just throwing a BadRequest in response to unexpected characters.