feat(parser): parse simple and block directives
Recursive Descent Parser calls on the Lexer for a token each time to
match based on the given grammar rules. This is only checking for the
validity of the syntax and no configurations are stored yet.
If an unexpected token is found, for now it simply throws an
UnexpectedToken(), but ideally this gets expanded into reporting precise
syntax errors.