feat(validator): implement validator that validates directives parsed
Unsure how well this will work out in terms of maintainability and
extensibility, but it's the simplest pattern I can think of right now,
so let's see how it goes.

A ConfigValidator will store a mapping of each Directive that we
support associated with a Validator - a function object that validates
the parameter of a Directive.

When we parse a Directive along with its Parameters, we not yet know
whether we should accept or reject that Parameter. The idea is to take
the Directive's name and look inside ConfigValidator, where if such
Directive cannot be found it means it is unsupported and an error should
be prompted. Where the Directive can be found, we call its' Validator on
the Parameter, checking that it fulfills the constraints of the Directive
otherwise we prompt the user with the acceptable values.