feat(parser): parse parameters and directives in block
A flag is added to Lexer to indicate whether we are now looking for a
Name or a Parameter. Each Parameter now gets stored inside a vector of
Strings, and our Validators can check the size of the vector to make
sure that the number of parameters are valid.

Directives in blocks are now correctly parsed until none are left. One
small difference I've noticed from Nginx is that if a block directive
was invalid (for example, being in the wrong context), it would
immediately be detected and errored. The way our parser is right now,
block directives aren't fully constructed until after everything within
is validated, so errors within the block would get reported before the
error of the block itself.