refactor(cookies): remove Session class
Remove storing and the use of Session class, the server will now
validate the necessary cookies (e.g. lang) directly rather than
first going through the session id.

At the most basic level the web server should not be storing any
information about the cookies, even if they do, it would theoretically
lead to poor server performance as it has to periodically check
for the session activity, expiry and so on.