feat(parser): store parsed directives inside a Configuration
Contexts are now represented using an enum rather than a String, and a
stack of Contexts indicate to us the surrounding contexts as we parse a
Directive.

Directives can be simple or blocks, where blocks can have "children",
which are simply just more directives. In my mind, this kind of
represents a treelike structure such as an AST, but I refuse to touch heap
memory myself, so I'll just stick with the std containers.
A Directive is now expanded to contain a vector of Directives,
which is appended to when parsing the Directives within a block.
Is it efficient that each time I return a new Directive B to be appended
to Directive A, it has to copy all of Directive B's childrens, which
could in turn contain more childrens? I'm not sure, but I would like to
not deal with memory leaks.