fix(parser): parse parameters for a block directive
Before this change, it was incorrectly deciding whether we are parsing a
simple directive or a block directive by looking at whether there is a
LCURLY after a Name. Now, we consume all parameters until after the
Name, until we hit a SEMICOLON or a LCURLY, which tells us whether this
is a simple or block directive.