refactor(validator): simplify validators to only throw and not return
Validators have no need to return anything, as their sole purpose is to
run through a bunch of checks for each Directive to ensure that it is
valid. Upon hitting an invalid Directive, there is nothing it can do but
throw an exception signalling the error for the caller of the Validator
to catch and report.