feat(parser): utilize multimap instead of map, stack to track mappings
A multimap of Directives is needed instead of a map because there are
Directives that may be duplicated within the same context. Now that
duplicates are in the picture, we also need a way to validate each
Directive's frequency in the same context, a stack of multimaps is used,
where each multimap contains the elements of its context. While
validating, we look at the top multimap of the stack and validate
against that scope. This does not feel elegant because there are a lot
of redundancies littered all over the place, ideally refactor later.