feat(servers): poll from multiple servers
Initially there was only a single socket listening on a hardcoded port
for incoming connections, now we have to spawn multiple servers
dynamically that are listening on potentially different ports. All the
polling and handling logic has been moved outside of a "server", and
will reside at a higher level that manages and oversees all the servers
which I have hastily named as "Servers", but I'm sure a semantically
better name can be chosen.