feat(Socket): encapsulate a Socket
A Socket object stores the fd to the socket and its underlying address.
Addresses are converted to their String representations only once and
saved within the object, so it doesn't have to be computed every call.
All the socket error handling can be done within the class, hidden away
from the general flow and we provide wrappers to access them.
A map keeps track of all the currently established socket connections.
This helps us work towards making sure there are no fd leaks as we can
have only one object that owns the Sockets, and have it close its fd on
destruction.