fix(directives): dynamically allocate Directives
While reading from a Configuration and applying its Directives into the
Server, it requires surrounding context from other Directives, which
requires each Directive to store a pointer back to its surrounding
environment. However, by using only stack allocation, it was (almost)
impossible to capture the surrounding state as there is no single
ownership and the Directives were constantly copied over and over as it
passes through functions and into containers.

Now, Directives are allocated once on the heap and pointers are passed
around and stored in multimaps. The Configuration is responsible for
cleaning up all the Directives when the Configuration is not needed
anymore.