refactor(GET): restructure the flow and remove unnecessary checks
Handle GET requests in priority of uploads, directories and normal
files.

Uploads for each session generate a listing if the directory itself is
requested, or return the file for that particular session.

Directories try with index files, returning the file to GET from the
filesystem, or generate a listing.

Retrieve regular files as normal.